+++ AllMyLights/Models/SinkOptions.cs
@@ -6,7 +6,10 @@ using Newtonsoft.Json;
 namespace AllMyLights.Models
 {
     [JsonConverter(typeof(InheritanceConverter), "Type", nameof(SinkOptions))]
-    [KnownType(typeof(OpenRGBSinkOptions))]
+    [
+        KnownType(typeof(OpenRGBSinkOptions)),
+        KnownType(typeof(WallpaperSinkOptions)),
+    ]
     public abstract class SinkOptions : ConnectorOptions
     {
     }