+++ AllMyLights/Models/WallpaperSinkOptions.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace AllMyLights.Models.OpenRGB
+{
+    public class WallpaperSinkOptions: SinkOptions
+    {
+        public string RelativeTo { get; set; }
+    }
+}
+
+