+++ AllMyLights/Platforms/IDesktop.cs
@@ -1,0 +1,7 @@
+﻿namespace AllMyLights.Platforms
+{
+    public interface IDesktop
+    {
+        void SetBackground(string filePath);
+    }
+}